pythonclient/requirements.txt
Flask==3.1
flask-cors==5.0.1
paho-mqtt==2.1.0
